Moved components to more appropriate folder Added .env sample file in ui directory
