+++ ui/.env.sample
@@ -1,0 +1,1 @@
+API_BASEURL=http://localhost:7879/
