+++ ui/src/config/index.ts
@@ -1,0 +1,3 @@
+export const API_BASEURL = import.meta.env.API_BASEURL;
+
+console.log('envs', import.meta.env);
